Move indexing API and document loader APIs to core
