+++ langchain-core/package.json
@@ -158,6 +158,15 @@
       "import": "./documents.js",
       "require": "./documents.cjs"
     },
+    "./document_loaders/base": {
+      "types": {
+        "import": "./document_loaders/base.d.ts",
+        "require": "./document_loaders/base.d.cts",
+        "default": "./document_loaders/base.d.ts"
+      },
+      "import": "./document_loaders/base.js",
+      "require": "./document_loaders/base.cjs"
+    },
     "./embeddings": {
       "types": {
         "import": "./embeddings.d.ts",
@@ -175,6 +184,15 @@
       },
       "import": "./example_selectors.js",
       "require": "./example_selectors.cjs"
+    },
+    "./indexing": {
+      "types": {
+        "import": "./indexing.d.ts",
+        "require": "./indexing.d.cts",
+        "default": "./indexing.d.ts"
+      },
+      "import": "./indexing.js",
+      "require": "./indexing.cjs"
     },
     "./language_models/base": {
       "types": {
@@ -595,6 +613,10 @@
     "documents.js",
     "documents.d.ts",
     "documents.d.cts",
+    "document_loaders/base.cjs",
+    "document_loaders/base.js",
+    "document_loaders/base.d.ts",
+    "document_loaders/base.d.cts",
     "embeddings.cjs",
     "embeddings.js",
     "embeddings.d.ts",
@@ -603,6 +625,10 @@
     "example_selectors.js",
     "example_selectors.d.ts",
     "example_selectors.d.cts",
+    "indexing.cjs",
+    "indexing.js",
+    "indexing.d.ts",
+    "indexing.d.cts",
     "language_models/base.cjs",
     "language_models/base.js",
     "language_models/base.d.ts",
