+++ langchain-core/src/indexing/index.ts
@@ -1,0 +1,2 @@
+export * from "./record_manager.js";
+export * from "./base.js";
