+++ langchain/src/indexes/index.ts
@@ -1,3 +1,14 @@
+import {
+  type CleanupMode,
+  type IndexOptions,
+  index,
+  _batch,
+  _deduplicateInOrder,
+  _getSourceIdAssigner,
+  _isBaseDocumentLoader,
+  _HashedDocument,
+} from "@langchain/core/indexing";
+
 export {
   type CleanupMode,
   type IndexOptions,
@@ -7,4 +18,4 @@ export {
   _getSourceIdAssigner,
   _isBaseDocumentLoader,
   _HashedDocument,
-} from "./indexing.js";
+};
